compatibility with examples; added multi-knapsack problem and added tests
