+++ doc/src/smoothed.rst
@@ -1,0 +1,12 @@
+.. _Smoothed:
+
+Smoothed
+========
+
+We consider various combinations of distribution estimation and resampling (bootstrap and bagging) for obtaining samples used to acquire a solution and for computing a confidence interval for an optimality gap, see paper "Distributions and Bootstrap for Data-based Stochastic Programming".
+
+Notes
+-----
+
+- Both "smoothed_B_I" and "smoothed_center_sample_size" are required in the json options even though one of them is needed.
+
