+++ examples/cvar/smoothed_cvar.json
@@ -1,5 +1,5 @@
 {
-    "module_name": "farmer",
+    "module_name": "cvar",
     "max_count": 8000, 
     "candidate_sample_size": 2,
     "sample_size": 20,
@@ -9,13 +9,10 @@
     "nB": 5,
     "alpha": 0.05,
     "seed_offset" :11,
-    "optimal_fname": "farmer_optimal.npy",
-    "xhat_fname": "farmer_xhat.npy",
+    "optimal_fname": "cvar_optimal.npy",
+    "xhat_fname": "cvar_xhat.npy",
     "solver_name": "gurobi_persistent",
     "trace_fname" : "resample_trace.txt",
     "boot_method": "Smoothed_bagging",
-    "coverage_replications": 10,
-    "crops_multiplier": 1,
-    "farmer_with_integers": "False",
-    "yield_cv": "0.1"
+    "coverage_replications": 10
 }
