+++ examples/multi_knapsack/multi_knapsack.json
@@ -1,5 +1,5 @@
 {
-    "module_name": "farmer",
+    "module_name": "multi_knapsack",
     "max_count": 8000, 
     "candidate_sample_size": 2,
     "sample_size": 20,
@@ -9,13 +9,11 @@
     "nB": 5,
     "alpha": 0.05,
     "seed_offset" :11,
-    "optimal_fname": "farmer_optimal.npy",
-    "xhat_fname": "farmer_xhat.npy",
+    "optimal_fname": "multi_knapsack_optimal.npy",
+    "xhat_fname": "multi_knapsack_xhat.npy",
     "solver_name": "gurobi_persistent",
     "trace_fname" : "resample_trace.txt",
     "boot_method": "Smoothed_bagging",
-    "coverage_replications": 10,
-    "crops_multiplier": 1,
-    "farmer_with_integers": "False",
-    "yield_cv": "0.1"
+    "coverage_replications": 800,
+    "deterministic_data_json": "multi_knapsack_data.json"
 }
